Insert all passed moves on insert schedule postpone to after implement log4net
by [aripmav1896@gmail.com] in [2011-10-11 15:12:42 (-03:00)] - hg hash {212e1e0e99fe}

+++ DFM.MVC/Areas/Accounts/Controllers/ReportController.cs
@@ -28,7 +28,7 @@ namespace DFM.MVC.Areas.Accounts.Controllers
 
 
 
-        public ActionResult SeeMonth(Int16? id)
+        public ActionResult SeeMonth(Int32? id)
         {
             var currentMonth = (Int16)DateTime.Today.Month;
